Try to fix issues with replay animations time
