+++ frontend/task/index.ts
@@ -53,7 +53,7 @@ import {PlayerInstant} from "../player";
 import {ActionTypes as StepperActionTypes, stepperClearError, stepperDisplayError} from "../stepper/actionTypes";
 import {ActionTypes as BufferActionTypes} from "../buffers/actionTypes";
 import {clearSourceHighlightSaga, StepperState, StepperStatus, StepperStepMode} from "../stepper";
-import {createQuickAlgoLibraryExecutor, StepperContext} from "../stepper/api";
+import {createQuickAlgoLibraryExecutor, makeContext, StepperContext} from "../stepper/api";
 import {taskSubmissionExecutor} from "./task_submission";
 import {ActionTypes as AppActionTypes} from "../actionTypes";
 import {ActionTypes as PlayerActionTypes} from "../player/actionTypes";
@@ -254,7 +254,7 @@ function* taskLoadSaga(app: App, action) {
 }
 
 function* handleLibrariesEventListenerSaga(app: App) {
-    const stepperContext: StepperContext = {
+    const stepperContext: StepperContext = makeContext(null, {
         interactAfter: (arg) => {
             return new Promise((resolve, reject) => {
                 app.dispatch({
@@ -264,11 +264,10 @@ function* handleLibrariesEventListenerSaga(app: App) {
                 });
             });
         },
+        environment: app.environment,
         dispatch: app.dispatch,
         quickAlgoContext: quickAlgoLibraries.getContext(null, app.environment),
-    };
-
-    stepperContext.quickAlgoCallsExecutor = createQuickAlgoLibraryExecutor(stepperContext);
+    });
 
     const listeners = quickAlgoLibraries.getEventListeners();
     log.getLogger('task').debug('create task listeners on ', app.environment, listeners);
